+++ packages/playwright-core/src/server/recorder/DEPS.list
@@ -7,6 +7,3 @@
 ../../utils/**
 ../../utilsBundle.ts
 ../../zipBundle.ts
-
-[recorderApp.ts]
-../chromium/crApp.ts
