+++ packages/playwright-core/src/server/trace/viewer/DEPS.list
@@ -4,4 +4,3 @@
 ../../../generated/
 ../../../utils/
 ../../../utilsBundle.ts
-../../chromium/crApp.ts
